Refactor codes and restrict access modifiers
The commit includes code refactoring for the SDK infrastructure of SwedbankPay, with modifications mainly in the PaymentOrder namespace. Some public records are changed to internal to restrict the access level. Also, several unnecessary using directives are removed and setters in DeviceDto class are changed to init only. The refactoring makes the code cleaner and improves encapsulation.