Refactor transaction request and reversal process
This commit refactors the transaction request and reversal process of a payment order. A TransactionRequest is replaced with a more specific PaymentOrderReversalTransaction class and PaymentOrderReversalTransactionDto. This change aims to make the code clearer, with each class and its methods having one specific use case, and to streamline the reversal process.